consolidate minerva tests
